+++ memo/D1351.history-timestamp.bash
@@ -10,5 +10,8 @@ shopt -s histappend
 source out/ble.sh
 
 #bleopt history_lazyload=
-bleopt history_share=1
-ble-attach
+#ble-attach
+
+#bleopt history_share=1
+
+blehook PRECMD+='builtin history -a'